+++ internal/sdkprovider/service/valkey/valkey_user.go
@@ -85,7 +85,7 @@ func ResourceValkeyUser() *schema.Resource {
 		CreateContext: common.WithGenClient(resourceValkeyUserCreate),
 		UpdateContext: common.WithGenClient(resourceValkeyUserUpdate),
 		ReadContext:   common.WithGenClient(resourceValkeyUserRead),
-		DeleteContext: common.WithGenClient(schemautil.DeleteResourceServiceUser),
+		DeleteContext: common.WithGenClient(schemautil.ResourceServiceUserDelete),
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
 		},